fix(desktop): avoid shell config for CLI command
